Fix SIMARM64 build on Mac.
R=zra@google.com

Review URL: https://codereview.chromium.org//243433005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@35189 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/simulator_arm64.cc
@@ -247,7 +247,7 @@ void SimulatorDebugger::Debug() {
             end = start + (10 * Instr::kInstrSize);
           }
         } else {
-          intptr_t length;
+          int64_t length;
           if (GetValue(arg1, &start) && GetValue(arg2, &length)) {
             if (Simulator::IsIllegalAddress(start)) {
               // If start isn't a valid address, warn and use PC instead
